+++ ui/src/components/expansion-item/QasExpansionItem.vue
@@ -1,13 +1,13 @@
 <template>
-  <div class="qas-expansion-item" :class="errorClasses">
-    <qas-box class="qas-expansion-item__box">
+  <div ref="expansionItem" class="qas-expansion-item" :class="errorClasses">
+    <component :is="component.is" class="qas-expansion-item__box">
       <q-expansion-item header-class="text-bold q-mt-sm q-pa-none" :label="props.label">
         <template #header>
           <slot name="header">
             <div class="full-width">
               <div class="items-center q-col-gutter-sm row">
                 <slot name="label">
-                  <h5 class="col-auto text-h5 text-weight-medium">
+                  <h5 class="col-auto text-h5">
                     {{ props.label }}
                   </h5>
                 </slot>
@@ -18,17 +18,23 @@
                   </div>
                 </div>
               </div>
+
+              <div v-if="hasHeaderBottom" class="q-mt-sm">
+                <slot name="header-bottom" />
+              </div>
             </div>
           </slot>
         </template>
 
-        <q-separator class="q-my-md" />
+        <q-separator v-if="!isNestedExpansionItem" class="q-my-md" />
 
         <slot name="content">
           <qas-grid-generator v-if="hasGridGenerator" v-bind="gridGeneratorProps" use-inline />
         </slot>
+
+        <q-separator v-if="hasBottomSeparator" class="q-mt-md" />
       </q-expansion-item>
-    </qas-box>
+    </component>
 
     <div v-if="hasError" class="q-pt-sm qas-expansion-item__error-message text-caption text-negative">
       {{ props.errorMessage }}
@@ -37,7 +43,9 @@
 </template>
 
 <script setup>
-import { computed } from 'vue'
+import QasBox from '../box/QasBox.vue'
+
+import { computed, provide, inject, onMounted, ref } from 'vue'
 
 defineOptions({ name: 'QasExpansionItem' })
 
@@ -67,16 +75,54 @@ const props = defineProps({
   }
 })
 
-const hasError = computed(() => props.error || !!props.errorMessage)
+provide('isExpansionItem', true)
 
+// slots
+const slots = defineSlots()
+
+// refs
+const expansionItem = ref(null)
+const hasNextSibling = ref(false)
+
+onMounted(setHasNextSibling)
+
+// constants
+const isNestedExpansionItem = inject('isExpansionItem', false)
+const component = {
+  is: isNestedExpansionItem ? 'div' : QasBox
+}
+
+// computed
+const hasError = computed(() => props.error || !!props.errorMessage)
 const errorClasses = computed(() => ({ 'qas-expansion-item--error': hasError.value }))
 
 const hasGridGenerator = computed(() => !!Object.keys(props.gridGeneratorProps).length)
+const hasBottomSeparator = computed(() => isNestedExpansionItem && hasNextSibling.value)
+const hasHeaderBottom = computed(() => !!slots['header-bottom'])
+
+// functions
+
+/**
+ * Caso o componente esteja dentro de um QasExpansionItem, verifica se existe um próximo irmão
+ * para adicionar um separador.
+ */
+function setHasNextSibling (value) {
+  if (!isNestedExpansionItem) return
+
+  const hasTextContentSibling = !!expansionItem.value.nextSibling.textContent?.trim?.()
+  const hasElementSibling = !!expansionItem.value.nextElementSibling
+
+  hasNextSibling.value = hasElementSibling || hasTextContentSibling
+}
 </script>
 
 <style lang="scss">
 .qas-expansion-item {
   $root: &;
+
+  & + & {
+    margin-top: var(--qas-spacing-lg);
+  }
 
   &--error {
     #{$root}__box {